requirements.txt
requests
beautifulsoup4
playwright
pyyaml
aiohttp 